+++ NEWS.md
@@ -2,7 +2,8 @@ writeAlizer v1.6.0 (Release date: 2025-9-10)
 ==============
 Changes: 
 * Replaced the RB mod3 models with their v2 counterparts (e.g., rb_mod3all now calls the rb_mod3all_v2 models).
-* Optimized code for the downloading of model objects.
+* Optimized code for the downloading of model objects, adding checksum verification.
+* predict_quality now outputs scores for submodels as well as the overall mean across submodels
 
 writeAlizer v1.5.0 (Release date: 2024-2-09)
 ==============